docs: Check ``stderr` with ``supervisorctl`` (#207)
I spend the afternoon yesterday to debug a particular problem with a service and couldn't see any error message. I contacted uberspace support (ticket [Uberspace.de #480399]) and Nico was so kind to email me back how to check ``stderr`` logs with ``supervisorctl``. Hope this helps other developers in the future.