Update web-domains.rst
I really had trouble to set up a subdomain and had to mail several times with the support. The docs are not sufficient, therefore I added a chapter with a tutorial containing all steps for setting up a subdomain.
Please double check if this is right and if my examples (e.g. for paths) are compatible with best practices.