Improved filter to match previous behaviour better
The current sieve filter only works for mail directly addressed mail. If the address isn't in one of the headers caught by ``address :matches "to"`` it doesn't work. This is the case for redirected mails and many mailing lists. Under the previous qmail subaddress system, those messages were caught.

Using the ``delivered-to`` header all mail delivered to subaddresses are delivered as before.

Minor changes:

* Condition with "anyof" to allow for multiple subaddresses
* Added comment regarding discard as alternative to reject